variants: refactoring and cleanup
